initialize the header validators
